Add feature `uniffi-binding` to restrict importing UniFFI dependencies (only necessary for Android and iOS).
