+++ common/build.rs
@@ -1,3 +1,4 @@
 fn main() {
+    #[cfg(feature = "uniffi-binding")]
     uniffi_build::generate_scaffolding("./src/common.udl").unwrap();
 }
